+++ VKClientRM/VKClientRM/Controllers/LoginViewController.swift
@@ -66,7 +66,7 @@ final class LoginViewController: UIViewController {
 
     // MARK: - Private Methods
 
-    @objc private func keyboardWillShown(notification: Notification) {
+    @objc private func keyboardWillShownAction(notification: Notification) {
         guard let info = notification.userInfo as? NSDictionary,
               let keyboard = info.value(forKey: UIResponder.keyboardFrameEndUserInfoKey) as? NSValue
         else { return }
@@ -75,14 +75,15 @@ final class LoginViewController: UIViewController {
         mainScrollView.scrollIndicatorInsets = contectInsets
     }
 
-    @objc private func keyboardWillHide(notification: Notification) {
+    @objc private func keyboardWillHideAction(notification: Notification) {
         mainScrollView.contentInset = UIEdgeInsets.zero
         mainScrollView.scrollIndicatorInsets = UIEdgeInsets.zero
     }
 
-    @objc private func hideKeyboard() {
+    @objc private func keyboardHideTapAction() {
         mainScrollView.endEditing(true)
     }
+
 
     private func setupView() {
         setupMainView()
@@ -140,19 +141,22 @@ final class LoginViewController: UIViewController {
     private func keyboardManager() {
         NotificationCenter.default.addObserver(
             self,
-            selector: #selector(keyboardWillShown(notification:)),
+            selector: #selector(keyboardWillShownAction(notification:)),
             name: UIResponder.keyboardWillShowNotification,
             object: nil
         )
         NotificationCenter.default.addObserver(
             self,
-            selector: #selector(keyboardWillHide(notification:)),
+            selector: #selector(keyboardWillHideAction(notification:)),
             name: UIResponder.keyboardWillHideNotification,
             object: nil
         )
-        let tapGesture = UITapGestureRecognizer(target: self, action: #selector(hideKeyboard))
-        mainScrollView.addGestureRecognizer(tapGesture)
     }
+
+    private func addTapGesture() {
+          let hideKeyboardGesture = UITapGestureRecognizer(target: self, action: #selector(keyboardHideTapAction))
+          mainScrollView.addGestureRecognizer(hideKeyboardGesture)
+      }
 
     private func removeKeyboardObserver() {
         NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillHideNotification, object: nil)
@@ -169,20 +173,3 @@ extension LoginViewController: UITextFieldDelegate {
     }
 }
 
-// MARK: - Alert
-
-extension UIViewController {
-    // MARK: - Constants
-
-    private enum Constants {
-        static let okText = "OK"
-    }
-
-    // MARK: - Public Methods
-
-    func showAlert(title: String, message: String) {
-        let alertController = UIAlertController(title: title, message: message, preferredStyle: .alert)
-        alertController.addAction(UIAlertAction(title: Constants.okText, style: .cancel, handler: nil))
-        present(alertController, animated: true)
-    }
-}
